# Improve perf of SolidPolygonLayer