--- a/src/experimental-layers/src/solid-polygon-layer/solid-polygon-layer-vertex-64.glsl.js
+++ b/src/experimental-layers/src/solid-polygon-layer/solid-polygon-layer-vertex-64.glsl.js
@@ -21,12 +21,16 @@
 export default `\
 #define SHADER_NAME solid-polygon-layer-vertex-shader-64
 
+attribute vec2 vertexPositions;
 attribute vec3 positions;
 attribute vec2 positions64xyLow;
-attribute vec3 normals;
+attribute vec3 nextPositions;
+attribute vec2 nextPositions64xyLow;
+attribute float elevations;
 attribute vec4 colors;
 attribute vec3 pickingColors;
 
+uniform float isSideVertex;
 uniform float extruded;
 uniform float elevationScale;
 uniform float opacity;
@@ -35,7 +39,23 @@ uniform vec3 pixelsPerUnit;
 varying vec4 vColor;
 
 void main(void) {
-  vec4 positions64xy = vec4(positions.x, positions64xyLow.x, positions.y, positions64xyLow.y);
+
+  vec3 pos;
+  vec2 pos64xyLow;
+  vec3 normal;
+
+  if (isSideVertex > 0.5) {
+    pos = mix(positions, nextPositions, vertexPositions.x);
+    pos64xyLow = mix(positions64xyLow, nextPositions64xyLow, vertexPositions.x);
+  } else {
+    pos = positions;
+    pos64xyLow = positions64xyLow;
+  }
+  if (extruded > 0.5) {
+    pos.z += elevations * vertexPositions.y;
+  }
+
+  vec4 positions64xy = vec4(pos.x, pos64xyLow.x, pos.y, pos64xyLow.y);
 
   vec2 projected_coord_xy[2];
   project_position_fp64(positions64xy, projected_coord_xy);
@@ -43,7 +63,7 @@ void main(void) {
   vec2 vertex_pos_modelspace[4];
   vertex_pos_modelspace[0] = projected_coord_xy[0];
   vertex_pos_modelspace[1] = projected_coord_xy[1];
-  vertex_pos_modelspace[2] = vec2(project_scale(positions.z * elevationScale), 0.0);
+  vertex_pos_modelspace[2] = vec2(project_scale(pos.z * elevationScale), 0.0);
   vertex_pos_modelspace[3] = vec2(1.0, 0.0);
 
   gl_Position = project_to_clipspace_fp64(vertex_pos_modelspace);
@@ -55,10 +75,14 @@ void main(void) {
   float lightWeight = 1.0;
 
   if (extruded > 0.5) {
-    lightWeight = getLightWeight(
-      position_worldspace.xyz, // the w component is always 1.0
-      normalize(normals * pixelsPerUnit)
-    );
+    if (isSideVertex > 0.5) {
+      normal = vec3(positions.y - nextPositions.y, nextPositions.x - positions.x, 0.0);
+      normal = normalize(normal * pixelsPerUnit);
+    } else {
+      normal = vec3(0.0, 0.0, 1.0);
+    }
+
+    lightWeight = getLightWeight(position_worldspace.xyz, normal);
   }
 
   vec3 lightWeightedColor = lightWeight * colors.rgb;