… src/experimental-layers/src/solid-polygon-layer/polygon-tesselator-extruded.js
… src/experimental-layers/test/polygon-tesselation.spec.js