This is a big change to an existing core layer, and utilizes an under-tested feature (multi-model layer). More testing will be required.

- Remove `PolygonTesselatorExtruded` class
- `SolidPolygonLayer` now renders filled sides or wireframe with an instanced model

Before:

| Test | Result |
| ---- | ------ |
| SolidPolygonLayer#initialize (flat) | 36.0 iterations/s (0.28s) |
| SolidPolygonLayer#initialize (extruded) | 9.39 iterations/s (1.06s) |
| SolidPolygonLayer#initialize (wireframe) | 29.5 iterations/s (0.34s) |
| SolidPolygonLayer#initialize (flat,fp64) | 14.5 iterations/s (0.69s) |
| SolidPolygonLayer#initialize (extruded,fp64) | 9.80 iterations/s (0.10s) |
| SolidPolygonLayer#initialize (wireframe,fp64) | 17.6 iterations/s (0.57s) |


After:

| Test | Result |
| ---- | ------ |
| SolidPolygonLayer#initialize (flat) | 43.1 iterations/s (0.23s) |
| SolidPolygonLayer#initialize (extruded) | 41.2 iterations/s (0.24s) |
| SolidPolygonLayer#initialize (wireframe) | 43.9 iterations/s (0.23s) |
| SolidPolygonLayer#initialize (flat,fp64) | 26.4 iterations/s (0.38s) |
| SolidPolygonLayer#initialize (extruded,fp64) | 22.2 iterations/s (0.45s) |
| SolidPolygonLayer#initialize (wireframe,fp64) | 25.8 iterations/s (0.39s) |

Tested: node, browser, layer-browser, website examples