@ibgreen 2017-12-10
@Pessimistress This is a fantastic speed improvement that should have a tremendous impact on impacted applications.

Since we are getting close to beta, before I start reviewing, what do you think of putting this version of the polygon layer into experimental-layer's for now? Then we would have more time to think through adding multi model layers to our core catalog...